+++ src/lib/conditions/checkMarketingFlag.js
@@ -5,8 +5,13 @@ var cookieHelper = require('../../helpers/cookieHelper');
 module.exports = function(settings) {
 
   var extensionSettings = turbine.getExtensionSettings();
+
+  //sync cookies
+  if(_satellite.cookie.get(extensionSettings.cookieNameServer) !== undefined){
+    _satellite.cookie.set(extensionSettings.cookieName, _satellite.cookie.get(extensionSettings.cookieNameServer),{ expires: 365, domain:'.css.ch', path:'/', SameSite:'Lax',secure:true });
+  } 
   
-  if(typeof _satellite.cookie.get(extensionSettings.cookieName) === 'undefined'){
+  if(typeof _satellite.cookie.get(extensionSettings.cookieName) === 'undefined' && typeof(_satellite.cookie.get(extensionSettings.cookieNameServer) === 'undefined')){
     
     return true;
   }