+++ releasenotes.md
@@ -1,9 +1,18 @@
 # Release Notes
 
+## Wedge v2.0.1
+- Bug fixing "init component"
+
+## Wedge v2.0.0
+- Added support for add consent in Switzerland
+- Added multiple actions and configurations for switzerland
+
+## Potato v1.5.5
+- Changed deprecated _satellite.buildInfo to _satellite.environment.stage
 ## Potato v1.4.2
-- Bug fixing serverside component
+- Bug fixing "serverside component"
 ## Potato v1.4.1
-- Bug fixing serverside component
+- Bug fixing "serverside component"
 ## Potato v1.4.0
 - Added a new UI Component to choose if cookie is set client- or serverside
 - Added serverside code for setting cookie via css cookie service
@@ -12,7 +21,7 @@
 ## Potato v1.3.1
 - Added UI for marketing and statistic tags
 ## Potato v1.3.0
-- Bug fixex
+- Bug fixes
 ## Potato v1.2.0 
 - Added Eventlistener for Cookie Settings
 - Added Rule for displaying Cookiebanner in CH